refactor: replace MCP server implementation with a new structure
- Removed the old MCP server implementation from `mcp-server.js`.
- Introduced a new `mcpApi.js` file that encapsulates the MCP server logic.
- Updated tool definitions and request handling methods for improved organization and maintainability.
- Enhanced error handling and logging for better debugging and user feedback.
- Maintained existing functionalities for product search, price comparison, and basket optimization.
